--- scripts/async-logger.cmake
@@ -1,12 +1,0 @@
-include(FetchContent)
-
-message("AsyncLogger")
-FetchContent_Declare(
-    AsyncLogger
-    GIT_REPOSITORY https://github.com/Yimura/AsyncLogger.git
-    GIT_TAG v0.0.6
-    GIT_PROGRESS TRUE
-)
-FetchContent_MakeAvailable(AsyncLogger)
-
-set_property(TARGET AsyncLogger PROPERTY CXX_STANDARD 23)
